# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  go_boilerplate_internal_modules_user_dto.AssignRoleRequest:
    properties:
      role_id:
        type: string
    required:
    - role_id
    type: object
  go_boilerplate_internal_modules_user_dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      password:
        maxLength: 50
        minLength: 6
        type: string
      role_id:
        description: 'Optional: if not provided, defaults to user role'
        type: string
    required:
    - email
    - name
    - password
    type: object
  go_boilerplate_internal_modules_user_dto.RoleInfo:
    properties:
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      slug:
        type: string
    type: object
  go_boilerplate_internal_modules_user_dto.UpdateUserRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 3
        type: string
      role_id:
        description: 'Optional: can update role to user or admin only'
        type: string
    type: object
  go_boilerplate_internal_modules_user_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
  go_boilerplate_internal_modules_user_dto.UserRoleResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      name:
        type: string
      role:
        $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.RoleInfo'
      updated_at:
        type: string
    type: object
  go_boilerplate_internal_shared_utils.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a robust Go REST API boilerplate using Fiber, GORM, and PostgreSQL.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /auth/me:
    get:
      description: Retrieve the profile information of the currently authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserRoleResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
...
  /users:
    get:
      description: Retrieve a paginated list of all registered users.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Users retrieved
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user account (Admin only).
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Create user'
      tags:
      - Users
  /users/{id}:
    delete:
      description: Delete a user account (Admin only).
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Delete user'
      tags:
      - Users
    get:
      description: Retrieve a user's basic profile information by their ID.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User retrieved
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user details. Users can update their own profile, or admins
        can update any user.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: Assign a specific role to a user account (Admin only).
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Role assignment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.AssignRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role assigned
          schema:
            allOf:
            - $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/go_boilerplate_internal_modules_user_dto.UserResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/go_boilerplate_internal_shared_utils.APIResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Assign role'
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"